Split the queue job into several files
