+++ src/queue/splitAudio/interfaces.ts
@@ -1,0 +1,7 @@
+export interface ISplitAudioInput {
+  /** The ID of the Transmission that we are splitting audio for. */
+  transmissionId: string;
+
+  /** The name of the Transmission that we are splitting audio for. */
+  transmissionName: string;
+}
